Modified vim's * behaviour
By default * jumps to the next occurence of the word under the cursor.
And by default * respects ignorecase and smartcase settings I have
configured. But I want * to not jump to the next occurence and have it
be case sensitive to show occurence of only that symbol in the current
buffer.